Move "globals" structs to the appropriate header files.
